+++ default.js
@@ -98,7 +98,7 @@ module.exports = {
     'no-lone-blocks': 'error',
     'no-loop-func': 'error',
     'no-magic-numbers': 'off',
-    'no-multi-spaces': 'error',
+    'no-multi-spaces': [ 'error', { 'ignoreEOLComments': true } ],
     'no-multi-str': 'error',
     'no-native-reassign': 'error',
     'no-new': 'error',