+++ CHANGELOG.md
@@ -11,6 +11,10 @@ details.
 
 ---
 
+## [Unreleased]
+### Added
+- checkrules.js: lookup multiple release tags in same package.
+
 ## [1.7.0] - 2017-03-05
 ### Added
 - All published rules as of ESLint 3.17.0